+++ android/src/main/kotlin/com/shopify/reactnative/flash_list/AutoLayoutView.kt
@@ -2,8 +2,6 @@ package com.shopify.reactnative.flash_list
 
 import android.content.Context
 import android.graphics.Canvas
-import android.util.DisplayMetrics
-import android.util.Log
 import android.view.View
 import android.view.ViewGroup
 import android.widget.HorizontalScrollView
@@ -21,8 +19,19 @@ class AutoLayoutView(context: Context) : ReactViewGroup(context) {
     val alShadow = AutoLayoutShadow()
     var enableInstrumentation = false
     var disableAutoLayout = false
+    var pixelDensity = 1.0
+    var experimentalScrollPositionManagement: Boolean = false
+        set(value: Boolean) {
+            alShadow.maintainTopContentPosition = value
+            field = value
+        }
 
-    var pixelDensity = 1.0;
+    override fun onLayout(changed: Boolean, left: Int, top: Int, right: Int, bottom: Int) {
+        if(experimentalScrollPositionManagement) {
+            fixLayout()
+        }
+        super.onLayout(changed, left, top, right, bottom)
+    }
 
     /** Overriding draw instead of onLayout. RecyclerListView uses absolute positions for each and every item which means that changes in child layouts may not trigger onLayout on this container. The same layout
      * can still cause views to overlap. Therefore, it makes sense to override draw to do correction. */
@@ -66,7 +75,10 @@ class AutoLayoutView(context: Context) : ReactViewGroup(context) {
             }
             positionSortedViews.sortBy { it.index }
             alShadow.offsetFromStart = if (alShadow.horizontal) left else top
-            alShadow.clearGapsAndOverlaps(positionSortedViews)
+            alShadow.clearGapsAndOverlaps(
+                    positionSortedViews,
+                    getParentScrollView() as ScrollView,
+            )
         }
     }