+++ android/src/main/kotlin/com/shopify/reactnative/flash_list/CellContainer.java
@@ -3,6 +3,8 @@ package com.shopify.reactnative.flash_list;
 public interface CellContainer {
     void setIndex(int value);
     int getIndex();
+    void setStableId(String value);
+    String getStableId();
     void setLeft(int value);
     int getLeft();
     void setTop(int value);