+++ android/src/main/kotlin/com/shopify/reactnative/flash_list/FlashListPackage.kt
@@ -13,7 +13,8 @@ class ReactNativeFlashListPackage : ReactPackage {
   override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
     return listOf(
         AutoLayoutViewManager(),
-        CellContainerManager()
+        CellContainerManager(),
+        DoubleSidedScrollViewManager()
     )
   }
 }