+++ ios/Sources/CellContainerManager.m
@@ -4,5 +4,6 @@
 @interface RCT_EXTERN_MODULE(CellContainerManager, RCTViewManager)
 
 RCT_EXPORT_VIEW_PROPERTY(index, NSInteger)
+RCT_EXPORT_VIEW_PROPERTY(stableId, NSString)
 
 @end