+++ src/native/auto-layout/AutoLayoutView.tsx
@@ -29,6 +29,7 @@ export interface AutoLayoutViewProps {
   onBlankAreaEvent?: BlankAreaEventHandler;
   onLayout?: (event: LayoutChangeEvent) => void;
   disableAutoLayout?: boolean;
+  experimentalScrollPositionManagement?: boolean;
 }
 
 class AutoLayoutView extends React.Component<AutoLayoutViewProps> {
@@ -63,6 +64,9 @@ class AutoLayoutView extends React.Component<AutoLayoutViewProps> {
           listeners.length !== 0 || Boolean(this.props.onBlankAreaEvent)
         }
         disableAutoLayout={this.props.disableAutoLayout}
+        experimentalScrollPositionManagement={Boolean(
+          this.props.experimentalScrollPositionManagement
+        )}
       >
         {this.props.children}
       </AutoLayoutViewNativeComponent>