+++ src/native/config/PlatformHelper.android.ts
@@ -4,6 +4,7 @@ const PlatformConfig = {
   defaultDrawDistance: 250,
   // Using rotate instead of scaleY on Android to avoid performance issues. Issue: https://github.com/Shopify/flash-list/issues/751
   invertedTransformStyle: { transform: [{ rotate: "180deg" }] },
+  invertedTransformStyleHorizontal: { transform: [{ rotate: "180deg" }] },
 };
 const getCellContainerPlatformStyles = (
   inverted: boolean,