+++ src/native/config/PlatformHelper.ios.ts
@@ -3,6 +3,7 @@ import { BaseItemAnimator } from "recyclerlistview";
 const PlatformConfig = {
   defaultDrawDistance: 250,
   invertedTransformStyle: { transform: [{ scaleY: -1 }] },
+  invertedTransformStyleHorizontal: { transform: [{ scaleX: -1 }] },
 };
 const getCellContainerPlatformStyles = (
   inverted: boolean,